+++ spec/graphql/types/iso_8601_date_spec.rb
@@ -51,10 +51,6 @@ describe GraphQL::Types::ISO8601Date do
 
     class Schema < GraphQL::Schema
       query(Query)
-      if TESTING_INTERPRETER
-        use GraphQL::Execution::Interpreter
-        use GraphQL::Analysis::AST
-      end
     end
   end