+++ spec/integration/rails/graphql/relay/range_add_spec.rb
@@ -51,7 +51,8 @@ describe GraphQL::Relay::RangeAdd do
       argument :menu_idx, Integer, required: true
 
       field :item_edge, item.edge_type, null: false
-      field :items, item.connection_type, null: false, connection: false
+      # On the old runtime, connection: false was required here:
+      field :items, item.connection_type, null: false, connection: TESTING_INTERPRETER
       field :menu, menu, null: false
 
       define_method :resolve do |input|
@@ -81,7 +82,17 @@ describe GraphQL::Relay::RangeAdd do
     Class.new(GraphQL::Schema) do
       self.query(query)
       self.mutation(mutation)
+      if TESTING_INTERPRETER
+        self.use(GraphQL::Analysis::AST)
+        self.use(GraphQL::Execution::Interpreter)
+      end
       self.cursor_encoder(PassThroughEncoder)
+
+      if TESTING_INTERPRETER
+        use GraphQL::Execution::Interpreter
+        use GraphQL::Analysis::AST
+        use GraphQL::Pagination::Connections
+      end
     end
   }